refactored calculator and interface code
