+++ EasyReflectometry/interfaces/__init__.py
@@ -4,23 +4,23 @@ __version__ = '0.0.1'
 import traceback
 
 try:
-    from EasyReflectometry.interfaces.refnx import Refnx  # noqa: F401
+    from .refnx.interface import Refnx  # noqa: F401
 except Exception:
     traceback.print_exc()
     print('Warning: refnx is not installed')
 
 # try:
-#     from EasyReflectometry.interfaces.bornagain import BornAgain  # noqa: F401
+#     from .bornagain import BornAgain  # noqa: F401
 # except Exception:
 #     traceback.print_exc()
 #     print('Warning: BornAgain python is not installed')
 
 try:
-    from EasyReflectometry.interfaces.refl1d import Refl1d  # noqa: F401
+    from .refl1d.interface import Refl1d  # noqa: F401
 except Exception:
     traceback.print_exc()
     print('Warning: refl1d is not installed')
 
-from EasyReflectometry.interfaces.interfaceTemplate import InterfaceTemplate
+from .interfaceTemplate import InterfaceTemplate
 
 _ = InterfaceTemplate
