+++ EasyReflectometry/interfaces/bornagain/interface.py
@@ -3,13 +3,14 @@ __author__ = "github.com/arm61"
 import numpy as np
 
 from easyCore.Objects.Inferface import ItemContainer
-from EasyReflectometry.interfaces.interfaceTemplate import InterfaceTemplate
-from EasyReflectometry.calculators.bornagain import BornAgain as BornAgain_calc
+
 from EasyReflectometry.sample.material import Material, MaterialMixture
 from EasyReflectometry.sample.layer import Layer
 from EasyReflectometry.sample.item import MultiLayer
 from EasyReflectometry.experiment.model import Model
 
+from ..interfaceTemplate import InterfaceTemplate
+from .wrapper import BornAgainWrapper
 
 class BornAgain(InterfaceTemplate):
     """
@@ -31,14 +32,14 @@ class BornAgain(InterfaceTemplate):
     name = 'BornAgain'
 
     def __init__(self):
-        self.calculator = BornAgain_calc()
+        self._wrapper = BornAgainWrapper()
         self._namespace = {}
 
     def reset_storage(self):
         """
         Reset the storage area of the calculator
         """
-        self.calculator.reset_storage()
+        self._wrapper.reset_storage()
 
     def create(self, model):
         """
@@ -53,42 +54,42 @@ class BornAgain(InterfaceTemplate):
         t_ = type(model)
         if issubclass(t_, Material):
             key = model.uid
-            if key not in self.calculator.storage['material'].keys():
-                self.calculator.create_material(key)
+            if key not in self._wrapper.storage['material'].keys():
+                self._wrapper.create_material(key)
             r_list.append(
                 ItemContainer(key, self._material_link,
-                              self.calculator.get_material_value,
-                              self.calculator.update_material))
+                              self._wrapper.get_material_value,
+                              self._wrapper.update_material))
         elif issubclass(t_, MaterialMixture):
             key = model.uid
-            if key not in self.calculator.storage['material'].keys():
-                self.calculator.create_material(key)
+            if key not in self._wrapper.storage['material'].keys():
+                self._wrapper.create_material(key)
             r_list.append(
                 ItemContainer(key, self._material_link,
-                              self.calculator.get_material_value,
-                              self.calculator.update_material))
+                              self._wrapper.get_material_value,
+                              self._wrapper.update_material))
         elif issubclass(t_, Layer):
             key = model.uid
-            if key not in self.calculator.storage['layer'].keys():
-                self.calculator.create_layer(key)
+            if key not in self._wrapper.storage['layer'].keys():
+                self._wrapper.create_layer(key)
             r_list.append(
-                ItemContainer(key, self._layer_link, self.calculator.get_layer_value,
-                              self.calculator.update_layer))
+                ItemContainer(key, self._layer_link, self._wrapper.get_layer_value,
+                              self._wrapper.update_layer))
             self.assign_material_to_layer(model.material.uid, key)
         elif issubclass(t_, MultiLayer):
             key = model.uid
-            self.calculator.create_item(key)
+            self._wrapper.create_item(key)
             r_list.append(
-                ItemContainer(key, self._item_link, self.calculator.get_item_value,
-                              self.calculator.update_item))
+                ItemContainer(key, self._item_link, self._wrapper.get_item_value,
+                              self._wrapper.update_item))
             for i in model.layers:
                 self.add_layer_to_item(i.uid, model.uid)
         elif issubclass(t_, Model):
-            self.calculator.create_model()
+            self._wrapper.create_model()
             r_list.append(
                 ItemContainer('model', self._model_link,
-                              self.calculator.get_model_value,
-                              self.calculator.update_model))
+                              self._wrapper.get_model_value,
+                              self._wrapper.update_model))
             for i in model.structure:
                 self.add_item_to_model(i.uid)
         return r_list
@@ -102,7 +103,7 @@ class BornAgain(InterfaceTemplate):
         :param layer_name: The layer name
         :type layer_name: str
         """
-        self.calculator.assign_material_to_layer(material_id, layer_id)
+        self._wrapper.assign_material_to_layer(material_id, layer_id)
 
     def add_layer_to_item(self, layer_id: int, item_id: int):
         """
@@ -113,7 +114,7 @@ class BornAgain(InterfaceTemplate):
         :param layer_id: The layer id
         :type layer_id: int
         """
-        self.calculator.add_layer_to_item(layer_id, item_id)
+        self._wrapper.add_layer_to_item(layer_id, item_id)
 
     def remove_layer_from_item(self, layer_id: int, item_id: int):
         """
@@ -124,7 +125,7 @@ class BornAgain(InterfaceTemplate):
         :param layer_id: The layer id
         :type layer_id: int
         """
-        self.calculator.remove_layer_from_item(layer_id, item_id)
+        self._wrapper.remove_layer_from_item(layer_id, item_id)
 
     def add_item_to_model(self, item_id: int):
         """
@@ -133,7 +134,7 @@ class BornAgain(InterfaceTemplate):
         :param item_id: The item id
         :type item_id: int
         """
-        self.calculator.add_item(item_id)
+        self._wrapper.add_item(item_id)
 
     def remove_item_from_model(self, item_id: int):
         """
@@ -144,7 +145,7 @@ class BornAgain(InterfaceTemplate):
         :param layer_id: The layer id
         :type layer_id: int
         """
-        self.calculator.remove_item(item_id)
+        self._wrapper.remove_item(item_id)
 
     def fit_func(self, x_array: np.ndarray) -> np.ndarray:
         """
@@ -154,7 +155,7 @@ class BornAgain(InterfaceTemplate):
         :return: calculated points
         :rtype: np.ndarray
         """
-        return self.calculator.calculate(x_array)
+        return self._wrapper.calculate(x_array)
 
     def sld_profile(self) -> tuple:
         """
@@ -163,4 +164,4 @@ class BornAgain(InterfaceTemplate):
         :return: z and sld(z)
         :rtype: tuple[np.ndarray, np.ndarray]
         """
-        return self.calculator.sld_profile()
+        return self._wrapper.sld_profile()
