+++ EasyReflectometry/interfaces/refl1d/interface.py
@@ -5,12 +5,13 @@ from typing import List, Tuple, Union
 import numpy as np
 
 from easyCore.Objects.Inferface import ItemContainer
-from EasyReflectometry.interfaces.interfaceTemplate import InterfaceTemplate
-from EasyReflectometry.calculators.refl1d import Refl1d as Refl1d_calc
 from EasyReflectometry.sample.material import Material, MaterialMixture
 from EasyReflectometry.sample.layer import Layer
 from EasyReflectometry.sample.item import MultiLayer
 from EasyReflectometry.experiment.model import Model
+
+from ..interfaceTemplate import InterfaceTemplate
+from .wrapper import Refl1dWrapper
 
 
 class Refl1d(InterfaceTemplate):
@@ -29,14 +30,14 @@ class Refl1d(InterfaceTemplate):
     name = 'refl1d'
 
     def __init__(self):
-        self.calculator = Refl1d_calc()
+        self._wrapper = Refl1dWrapper()
         self._namespace = {}
 
     def reset_storage(self):
         """
         Reset the storage area of the calculator
         """
-        self.calculator.reset_storage()
+        self._wrapper.reset_storage()
 
     def create(self, model: Union[Material, Layer, MultiLayer,
                                   Model]) -> List[ItemContainer]:
@@ -50,42 +51,42 @@ class Refl1d(InterfaceTemplate):
         t_ = type(model)
         if issubclass(t_, Material):
             key = model.uid
-            if key not in self.calculator.storage['material'].keys():
-                self.calculator.create_material(key)
+            if key not in self._wrapper.storage['material'].keys():
+                self._wrapper.create_material(key)
             r_list.append(
                 ItemContainer(key, self._material_link,
-                              self.calculator.get_material_value,
-                              self.calculator.update_material))
+                              self._wrapper.get_material_value,
+                              self._wrapper.update_material))
         elif issubclass(t_, MaterialMixture):
             key = model.uid
-            if key not in self.calculator.storage['material'].keys():
-                self.calculator.create_material(key)
+            if key not in self._wrapper.storage['material'].keys():
+                self._wrapper.create_material(key)
             r_list.append(
                 ItemContainer(key, self._material_link,
-                              self.calculator.get_material_value,
-                              self.calculator.update_material))
+                              self._wrapper.get_material_value,
+                              self._wrapper.update_material))
         elif issubclass(t_, Layer):
             key = model.uid
-            if key not in self.calculator.storage['layer'].keys():
-                self.calculator.create_layer(key)
+            if key not in self._wrapper.storage['layer'].keys():
+                self._wrapper.create_layer(key)
             r_list.append(
-                ItemContainer(key, self._layer_link, self.calculator.get_layer_value,
-                              self.calculator.update_layer))
+                ItemContainer(key, self._layer_link, self._wrapper.get_layer_value,
+                              self._wrapper.update_layer))
             self.assign_material_to_layer(model.material.uid, key)
         elif issubclass(t_, MultiLayer):
             key = model.uid
-            self.calculator.create_item(key)
+            self._wrapper.create_item(key)
             r_list.append(
-                ItemContainer(key, self._item_link, self.calculator.get_item_value,
-                              self.calculator.update_item))
+                ItemContainer(key, self._item_link, self._wrapper.get_item_value,
+                              self._wrapper.update_item))
             for i in model.layers:
                 self.add_layer_to_item(i.uid, model.uid)
         elif issubclass(t_, Model):
             key = model.uid
-            self.calculator.create_model(key)
+            self._wrapper.create_model(key)
             r_list.append(
-                ItemContainer(key, self._model_link, self.calculator.get_model_value,
-                              self.calculator.update_model))
+                ItemContainer(key, self._model_link, self._wrapper.get_model_value,
+                              self._wrapper.update_model))
             for i in model.structure:
                 self.add_item_to_model(i.uid, key)
         return r_list
@@ -97,7 +98,7 @@ class Refl1d(InterfaceTemplate):
         :param material_name: The material name
         :param layer_name: The layer name
         """
-        self.calculator.assign_material_to_layer(material_id, layer_id)
+        self._wrapper.assign_material_to_layer(material_id, layer_id)
 
     def add_layer_to_item(self, layer_id: str, item_id: str):
         """
@@ -106,7 +107,7 @@ class Refl1d(InterfaceTemplate):
         :param item_id: The item id
         :param layer_id: The layer id
         """
-        self.calculator.add_layer_to_item(layer_id, item_id)
+        self._wrapper.add_layer_to_item(layer_id, item_id)
 
     def remove_layer_from_item(self, layer_id: str, item_id: str):
         """
@@ -115,7 +116,7 @@ class Refl1d(InterfaceTemplate):
         :param item_id: The item id
         :type item_id: int        :param layer_id: The layer id
         """
-        self.calculator.remove_layer_from_item(layer_id, item_id)
+        self._wrapper.remove_layer_from_item(layer_id, item_id)
 
     def add_item_to_model(self, item_id: str, model_id: str):
         """
@@ -124,7 +125,7 @@ class Refl1d(InterfaceTemplate):
         :param item_id: The item id
         :param model_id: The model id
         """
-        self.calculator.add_item(item_id, model_id)
+        self._wrapper.add_item(item_id, model_id)
 
     def remove_item_from_model(self, item_id: str, model_id: str):
         """
@@ -133,7 +134,7 @@ class Refl1d(InterfaceTemplate):
         :param item_id: The item id
         :param model_id: The model id
         """
-        self.calculator.remove_item(item_id, model_id)
+        self._wrapper.remove_item(item_id, model_id)
 
     def fit_func(self, x_array: np.ndarray, model_id: str) -> np.ndarray:
         """
@@ -143,7 +144,7 @@ class Refl1d(InterfaceTemplate):
         :param model_id: The name of the model
         :return: calculated points
         """
-        return self.calculator.calculate(x_array, model_id)
+        return self._wrapper.calculate(x_array, model_id)
 
     def sld_profile(self, model_id: str) -> Tuple[np.ndarray, np.ndarray]:
         """
@@ -151,4 +152,4 @@ class Refl1d(InterfaceTemplate):
 
         :return: z and sld(z)
         """
-        return self.calculator.sld_profile(model_id)
+        return self._wrapper.sld_profile(model_id)
