+++ EasyReflectometry/interfaces/refnx/interface.py
@@ -5,12 +5,14 @@ from typing import List, Tuple, Union
 import numpy as np
 
 from easyCore.Objects.Inferface import ItemContainer
-from EasyReflectometry.interfaces.interfaceTemplate import InterfaceTemplate
-from EasyReflectometry.calculators.refnx import Refnx as Refnx_calc
+
 from EasyReflectometry.sample.material import Material, MaterialMixture
 from EasyReflectometry.sample.layer import Layer
 from EasyReflectometry.sample.item import MultiLayer
 from EasyReflectometry.experiment.model import Model
+
+from ..interfaceTemplate import InterfaceTemplate
+from .wrapper import RefnxWrapper
 
 
 class Refnx(InterfaceTemplate):
@@ -29,14 +31,14 @@ class Refnx(InterfaceTemplate):
     name = 'refnx'
 
     def __init__(self):
-        self.calculator = Refnx_calc()
+        self._wrapper = RefnxWrapper()
         self._namespace = {}
 
     def reset_storage(self):
         """
         Reset the storage area of the calculator
         """
-        self.calculator.reset_storage()
+        self._wrapper.reset_storage()
 
     def create(self, model: Union[Material, Layer, MultiLayer,
                                   Model]) -> List[ItemContainer]:
@@ -50,42 +52,42 @@ class Refnx(InterfaceTemplate):
         t_ = type(model)
         if issubclass(t_, Material):
             key = model.uid
-            if key not in self.calculator.storage['material'].keys():
-                self.calculator.create_material(key)
+            if key not in self._wrapper.storage['material'].keys():
+                self._wrapper.create_material(key)
             r_list.append(
                 ItemContainer(key, self._material_link,
-                              self.calculator.get_material_value,
-                              self.calculator.update_material))
+                              self._wrapper.get_material_value,
+                              self._wrapper.update_material))
         elif issubclass(t_, MaterialMixture):
             key = model.uid
-            if key not in self.calculator.storage['material'].keys():
-                self.calculator.create_material(key)
+            if key not in self._wrapper.storage['material'].keys():
+                self._wrapper.create_material(key)
             r_list.append(
                 ItemContainer(key, self._material_link,
-                              self.calculator.get_material_value,
-                              self.calculator.update_material))
+                              self._wrapper.get_material_value,
+                              self._wrapper.update_material))
         elif issubclass(t_, Layer):
             key = model.uid
-            if key not in self.calculator.storage['layer'].keys():
-                self.calculator.create_layer(key)
+            if key not in self._wrapper.storage['layer'].keys():
+                self._wrapper.create_layer(key)
             r_list.append(
-                ItemContainer(key, self._layer_link, self.calculator.get_layer_value,
-                              self.calculator.update_layer))
+                ItemContainer(key, self._layer_link, self._wrapper.get_layer_value,
+                              self._wrapper.update_layer))
             self.assign_material_to_layer(model.material.uid, key)
         elif issubclass(t_, MultiLayer):
             key = model.uid
-            self.calculator.create_item(key)
+            self._wrapper.create_item(key)
             r_list.append(
-                ItemContainer(key, self._item_link, self.calculator.get_item_value,
-                              self.calculator.update_item))
+                ItemContainer(key, self._item_link, self._wrapper.get_item_value,
+                              self._wrapper.update_item))
             for i in model.layers:
                 self.add_layer_to_item(i.uid, model.uid)
         elif issubclass(t_, Model):
             key = model.uid
-            self.calculator.create_model(key)
+            self._wrapper.create_model(key)
             r_list.append(
-                ItemContainer(key, self._model_link, self.calculator.get_model_value,
-                              self.calculator.update_model))
+                ItemContainer(key, self._model_link, self._wrapper.get_model_value,
+                              self._wrapper.update_model))
             for i in model.structure:
                 self.add_item_to_model(i.uid, key)
         return r_list
@@ -97,7 +99,7 @@ class Refnx(InterfaceTemplate):
         :param material_name: The material name
         :param layer_name: The layer name
         """
-        self.calculator.assign_material_to_layer(material_id, layer_id)
+        self._wrapper.assign_material_to_layer(material_id, layer_id)
 
     def add_layer_to_item(self, layer_id: str, item_id: str):
         """
@@ -106,7 +108,7 @@ class Refnx(InterfaceTemplate):
         :param item_id: The item id
         :param layer_id: The layer id
         """
-        self.calculator.add_layer_to_item(layer_id, item_id)
+        self._wrapper.add_layer_to_item(layer_id, item_id)
 
     def remove_layer_from_item(self, layer_id: str, item_id: str):
         """
@@ -115,7 +117,7 @@ class Refnx(InterfaceTemplate):
         :param item_id: The item id
         :param layer_id: The layer id
         """
-        self.calculator.remove_layer_from_item(layer_id, item_id)
+        self._wrapper.remove_layer_from_item(layer_id, item_id)
 
     def add_item_to_model(self, item_id: str, model_id: str):
         """
@@ -124,7 +126,7 @@ class Refnx(InterfaceTemplate):
         :param item_id: The item id
         :param model_id: The model id
         """
-        self.calculator.add_item(item_id, model_id)
+        self._wrapper.add_item(item_id, model_id)
 
     def remove_item_from_model(self, item_id: str, model_id: str):
         """
@@ -133,7 +135,7 @@ class Refnx(InterfaceTemplate):
         :param item_id: The item id
         :param model_id: The model id
         """
-        self.calculator.remove_item(item_id, model_id)
+        self._wrapper.remove_item(item_id, model_id)
 
     def change_item_to_repeating_multi_layer(self, item_id: str, old_id: str):
         """
@@ -142,7 +144,7 @@ class Refnx(InterfaceTemplate):
         :param item_name: The item name
         :param old_id: id of old item
         """
-        self.calculator.change_item_to_repeating_multi_layer(item_id, old_id)
+        self._wrapper.change_item_to_repeating_multi_layer(item_id, old_id)
 
     def fit_func(self, x_array: np.ndarray, model_id: str) -> np.ndarray:
         """
@@ -152,7 +154,7 @@ class Refnx(InterfaceTemplate):
         :param model_id: The model id
         :return: calculated points
         """
-        return self.calculator.calculate(x_array, model_id)
+        return self._wrapper.calculate(x_array, model_id)
 
     def sld_profile(self, model_id: str) -> Tuple[np.ndarray, np.ndarray]:
         """
@@ -161,4 +163,4 @@ class Refnx(InterfaceTemplate):
         :param model_id: The model id
         :return: z and sld(z)
         """
-        return self.calculator.sld_profile(model_id)
+        return self._wrapper.sld_profile(model_id)
