+++ tests/experiment/test_model.py
@@ -105,11 +105,11 @@ class TestModel(unittest.TestCase):
         o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
         d = Structure.from_pars(o1, name='myModel')
         mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
         mod.add_item(o2)
-        assert_equal(len(mod.interface().calculator.storage['item']), 2)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
 
     def test_add_item_with_interface_refl1d(self):
         interface = InterfaceFactory()
@@ -124,11 +124,11 @@ class TestModel(unittest.TestCase):
         o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
         d = Structure.from_pars(o1, name='myModel')
         mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
         mod.add_item(o2)
-        assert_equal(len(mod.interface().calculator.storage['item']), 2)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
 
     # def test_add_item_with_interface_bornagain(self):
     #     interface = InterfaceFactory()
@@ -143,11 +143,11 @@ class TestModel(unittest.TestCase):
     #     o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
     #     d = Structure.from_pars(o1, name='myModel')
     #     mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 1)
-    #     assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+    #     assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
     #     mod.add_item(o2)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 2)
-    #     assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
 
     def test_duplicate_item(self):
         m1 = Material.from_pars(6.908, -0.278, 'Boron')
@@ -180,11 +180,11 @@ class TestModel(unittest.TestCase):
         o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
         d = Structure.from_pars(o1, name='myModel')
         mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
         mod.add_item(o2)
-        assert_equal(len(mod.interface().calculator.storage['item']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
         mod.duplicate_item(1)
-        assert_equal(len(mod.interface().calculator.storage['item']), 3)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 3)
 
     def test_duplicate_item_with_interface_refl1d(self):
         interface = InterfaceFactory()
@@ -199,11 +199,11 @@ class TestModel(unittest.TestCase):
         o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
         d = Structure.from_pars(o1, name='myModel')
         mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
         mod.add_item(o2)
-        assert_equal(len(mod.interface().calculator.storage['item']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
         mod.duplicate_item(1)
-        assert_equal(len(mod.interface().calculator.storage['item']), 3)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 3)
 
     # def test_duplicate_item_with_interface_bornagain(self):
     #     interface = InterfaceFactory()
@@ -218,11 +218,11 @@ class TestModel(unittest.TestCase):
     #     o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
     #     d = Structure.from_pars(o1, name='myModel')
     #     mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 1)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
     #     mod.add_item(o2)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
     #     mod.duplicate_item(1)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 3)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 3)
 
     def test_remove_item(self):
         m1 = Material.from_pars(6.908, -0.278, 'Boron')
@@ -253,14 +253,14 @@ class TestModel(unittest.TestCase):
         o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
         d = Structure.from_pars(o1, name='myModel')
         mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
         mod.add_item(o2)
-        assert_equal(len(mod.interface().calculator.storage['item']), 2)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
         mod.remove_item(0)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
 
     def test_remove_item_with_interface_refl1d(self):
         interface = InterfaceFactory()
@@ -275,14 +275,14 @@ class TestModel(unittest.TestCase):
         o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
         d = Structure.from_pars(o1, name='myModel')
         mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
         mod.add_item(o2)
-        assert_equal(len(mod.interface().calculator.storage['item']), 2)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
         mod.remove_item(0)
-        assert_equal(len(mod.interface().calculator.storage['item']), 1)
-        assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+        assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+        assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
 
     # def test_remove_item_with_interface_bornagain(self):
     #     interface = InterfaceFactory()
@@ -297,14 +297,14 @@ class TestModel(unittest.TestCase):
     #     o2 = RepeatingMultiLayer.from_pars(ls2, 1.0, 'oneLayerItem2')
     #     d = Structure.from_pars(o1, name='myModel')
     #     mod = Model.from_pars(d, 2, 1e-5, 2.0, 'newModel', interface=interface)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 1)
-    #     assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+    #     assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
     #     mod.add_item(o2)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 2)
-    #     assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
     #     mod.remove_item(0)
-    #     assert_equal(len(mod.interface().calculator.storage['item']), 1)
-    #     assert_equal(len(mod.interface().calculator.storage['layer']), 2)
+    #     assert_equal(len(mod.interface()._wrapper.storage['item']), 1)
+    #     assert_equal(len(mod.interface()._wrapper.storage['layer']), 2)
 
     def test_uid(self):
         p = Model.default()
